requirements.txt
langchain-google-genai
langchain-core
langchain-community
streamlit
python-dotenv
langchain-ollama